Fix a panic when getting the list of voices
We should not subtract 1 from voice_num since the range does not include
the end value. This was causing the last voice not to be included in the
list and we were also occasionally seeing the following panic:

Error { message: "SynthesisVoicesResult::from_handle(synthesis_voices_result_get_voice_info) error: vector::_M_range_check: __n (which is 0) >= this->size() (which is 0)", caused_by: ApiError(3) }

I think this was caused by the Azure API returning 0 voices and then
voice_num was underflowing, resulting in calling synthesis_voices_result_get_voice_info
with an idx that is out of range (e.g. 4294967295). I'm not sure why
Azure would return 0 voices from the API, though.