Fix a panic when getting the list of voices (#26)
* Fix a panic when getting the list of voices

We should not subtract 1 from voice_num since the range does not include
the end value. This was causing the last voice not to be included in the
list and we were also occasionally seeing the following panic:

Error { message: "SynthesisVoicesResult::from_handle(synthesis_voices_result_get_voice_info) error: vector::_M_range_check: __n (which is 0) >= this->size() (which is 0)", caused_by: ApiError(3) }

I think this was caused by the Azure API returning 0 voices and then
voice_num was underflowing, resulting in calling synthesis_voices_result_get_voice_info
with an idx that is out of range (e.g. 4294967295). I'm not sure why
Azure would return 0 voices from the API, though.

* Fix dangling pointer warnings

```
warning: a dangling pointer will be produced because the temporary `CString` will be dropped
   --> ./cognitive-services-speech-sdk-rs/src/audio/push_audio_input_stream.rs:111:48
    |
111 |             let c_value = CString::new(value)?.as_ptr();
    |                           -------------------- ^^^^^^ this pointer will immediately be invalid
    |                           |
    |                           this `CString` is deallocated at the end of the statement, bind it to a variable to extend its lifetime
    |
    = note: pointers do not have a lifetime; when calling `as_ptr` the `CString` will be deallocated at the end of the statement because nothing is referencing it as far as the type system is concerned
    = help: you must make sure that the variable you bind the `CString` to lives at least as long as the pointer returned by the call to `as_ptr`
    = help: in particular, if this pointer is returned from the current function, binding the `CString` inside the function will not suffice
    = help: for more information, see <https://doc.rust-lang.org/reference/destructors.html>
    = note: `#[warn(dangling_pointers_from_temporaries)]` on by default
```

* Fix a typo